app/core/payment_processor.py
from typing import Dict, List
from web3 import Web3
from eth_typing import Address
import os
from dotenv import load_load_dotenv

load_dotenv()

class PaymentProcessor:
    def __init__(self):
        self.w3 = Web3(Web3.HTTPProvider(os.getenv("WEB3_PROVIDER_URI")))
        self.stablecoin_address = os.getenv("STABLECOIN_ADDRESS")
        self.contract_abi = self._load_contract_abi()
        
    def _load_contract_abi(self) -> List:
        """Load the stablecoin contract ABI"""
        # This would typically load from a JSON file
        return []  # Replace with actual ABI
    
    async def process_payments(self, distributions: Dict[str, float]) -> List[Dict]:
        """Process payments according to the distribution"""
        transactions = []
        
        for recipient, amount in distributions.items():
            tx = await self._send_payment(recipient, amount)
            transactions.append({
                "recipient": recipient,
                "amount": amount,
                "transaction_hash": tx
            })
        
        return transactions
    
    async def _send_payment(self, recipient: str, amount: float) -> str:
        """Send payment to a recipient"""
        # This is a placeholder for actual blockchain transaction logic
        # In a real implementation, this would:
        # 1. Convert amount to wei
        # 2. Create and sign transaction
        # 3. Send transaction to blockchain
        # 4. Return transaction hash
        return f"0x{os.urandom(32).hex()}"  # Placeholder transaction hash
