requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
web3==6.11.1
sqlalchemy==2.0.23
alembic==1.12.1
pytest==7.4.3
httpx==0.25.1
python-multipart==0.0.6
redis==5.0.1
celery==5.3.4
psycopg2-binary==2.9.9
fastapi-users[sqlalchemy]==12.1.2
aiohttp==3.9.1
eth-brownie==1.19.3
web3-ethereum-defi==0.1.4
asyncpg==0.29.0
fastapi-cache2==0.2.1
prometheus-client==0.19.0
sentry-sdk==1.32.0
